@@ -1,16 +1,16 @@
 {
-  "name": "metis-ui",
+  "name": "authkeeper-ui",
   "version": "0.1.0",
-  "description": "Metis for Testing, build with React + Vite + TypeScript and thanks for Vital.",
-  "homepage": "https://metis.tabuyos.com",
+  "description": "Auth Keeper for Testing, build with React + Vite + TypeScript and thanks for Vital.",
+  "homepage": "https://authkeeper.tabuyos.com",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/service-io/metis"
+    "url": "git+https://github.com/service-io/authkeeper"
   },
   "author": "Tabuyos <tabuyos@outlook.com>",
   "license": "BSD-3-Clause",
   "bugs": {
-    "url": "https://github.com/service-io/metis/issues"
+    "url": "https://github.com/service-io/authkeeper/issues"
   },
   "keywords": [
     "react",
@@ -32,47 +32,46 @@
     "not IE 11"
   ],
   "engines": {
-    "node": ">=14",
-    "yarn": ">=1.22.5"
+    "node": ">=16.20.0",
+    "yarn": ">=1.22.19"
   },
   "scripts": {
     "start": "vite",
     "build": "vite build",
-    "test": "jest",
-    "postinstall": "cd .. && husky install ui/.husky"
+    "test": "jest"
   },
   "dependencies": {
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "@heroicons/react": "2.0.13"
+    "react-router-dom": "6.15.0",
+    "@heroicons/react": "2.0.18"
   },
   "devDependencies": {
-    "@types/jest": "29.2.5",
-    "@types/node": "18.11.18",
-    "@types/react": "18.0.26",
-    "@typescript-eslint/eslint-plugin": "5.48.1",
-    "@typescript-eslint/parser": "5.48.1",
-    "vite": "4.0.4",
-    "vite-preset-react": "2.3.0",
-    "typescript": "4.9.4",
-    "postcss": "8.4.21",
-    "prettier": "2.8.2",
-    "lint-staged": "13.1.0",
-    "husky": "8.0.3",
-    "tailwindcss": "3.2.4",
-    "@tailwindcss/forms": "0.5.3",
-    "eslint": "8.31.0",
-    "eslint-plugin-import": "2.27.4",
-    "eslint-plugin-react": "7.32.0",
+    "@types/jest": "29.5.3",
+    "@types/node": "20.5.1",
+    "@types/react": "18.2.20",
+    "@typescript-eslint/eslint-plugin": "6.4.0",
+    "@typescript-eslint/parser": "6.4.0",
+    "vite": "4.4.9",
+    "@vitejs/plugin-react": "4.0.4",
+    "typescript": "5.1.6",
+    "postcss": "8.4.28",
+    "prettier": "3.0.2",
+    "lint-staged": "14.0.0",
+    "tailwindcss": "3.3.3",
+    "@tailwindcss/forms": "0.5.4",
+    "eslint": "8.47.0",
+    "eslint-plugin-import": "2.28.1",
+    "eslint-plugin-react": "7.33.2",
     "eslint-plugin-react-hooks": "4.6.0",
-    "autoprefixer": "10.4.13",
-    "@commitlint/cli": "17.4.2",
-    "@commitlint/config-conventional": "17.4.2",
-    "ts-jest": "29.0.5",
-    "jest": "29.3.1",
-    "@testing-library/react": "13.4.0",
-    "@testing-library/jest-dom": "5.16.5",
-    "jest-environment-jsdom": "29.3.1",
+    "autoprefixer": "10.4.15",
+    "@commitlint/cli": "17.7.1",
+    "@commitlint/config-conventional": "17.7.0",
+    "ts-jest": "29.1.1",
+    "jest": "29.6.2",
+    "@testing-library/react": "14.0.0",
+    "@testing-library/jest-dom": "6.0.1",
+    "jest-environment-jsdom": "29.6.2",
     "identity-obj-proxy": "3.0.0"
   }
 }
